refactor: 하위 뷰모델 생성
