+++ android/app/src/main/java/com/daedan/festabook/presentation/placeList/placeMap/MapClickListenerImpl.kt
@@ -1,0 +1,25 @@
+package com.daedan.festabook.presentation.placeList.placeMap
+
+import com.daedan.festabook.presentation.placeList.PlaceListViewModel
+import com.daedan.festabook.presentation.placeList.model.PlaceCategoryUiModel
+import timber.log.Timber
+
+class MapClickListenerImpl(
+    private val viewModel: PlaceListViewModel,
+) : MapClickListener {
+    override fun onMarkerListener(
+        placeId: Long,
+        category: PlaceCategoryUiModel,
+    ): Boolean {
+        Timber.d("Marker CLick : placeID: $placeId categoty: $category")
+        if (category !in PlaceCategoryUiModel.SECONDARY_CATEGORIES) {
+            viewModel.selectPlace(placeId, category)
+        }
+        return true
+    }
+
+    override fun onMapClickListener() {
+        Timber.d("Map CLick")
+        viewModel.unselectPlace()
+    }
+}
